requirements.txt
git+https://github.com/Shinyhero36/Leboncoin-API-Wrapper.git@483accb5b5eed0#egg=leboncoin_api_wrapper
docopt
requests
pyyaml
pydantic
